backend/requirements.txt
annotated-types>=0.7.0
asgiref>=3.11.0
cachetools>=6.2.2
certifi>=2025.11.12
charset-normalizer>=3.4.4
Django>=5.2.9
django-cors-headers>=4.9.0
djangorestframework>=3.16.1
google-ai-generativelanguage
google-api-core
google-api-python-client>=2.187.0
google-auth>=2.43.0
google-auth-httplib2>=0.2.1
google-generativeai>=0.8.5
googleapis-common-protos
grpcio>=1.76.0
grpcio-status>=1.71.2
httplib2>=0.31.0
idna>=3.11
proto-plus>=1.26.1
protobuf
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pydantic>=2.12.5
pydantic_core>=2.41.5
pyparsing>=3.2.5
requests>=2.32.5
rsa>=4.9.1
sqlparse>=0.5.4
tqdm>=4.67.1
typing-inspection>=0.4.2
typing_extensions>=4.15.0
uritemplate>=4.2.0
urllib3>=2.5.0
